new tile class for caching components
