+++ Assets/Resources/Scripts/Map/SaveHandler/TileLoader.cs
@@ -28,7 +28,7 @@ public class TileLoader : MonoBehaviour
             GameObject curTile = createTile(t);
             boardTiles[i, j] = curTile;
             tileLookup.Add(curTile, (i, j));
-            curTile.GetComponent<Tile>().id = j * width + i;
+            curTile.GetComponent<TileLocation>().id = j * width + i;
 
             i++;
             if (i >= width)
@@ -45,7 +45,7 @@ public class TileLoader : MonoBehaviour
     public GameObject createTile(SerializedTile st)
     {
         GameObject newTile = Instantiate(TileTemplate);
-        newTile.GetComponent<Tile>().board = curBoard.GetComponent<Board>();
+        newTile.GetComponent<TileLocation>().board = curBoard.GetComponent<Board>();
 
         JsonUtility.FromJsonOverwrite(st.serializedComponents[0], newTile.GetComponent<TileChars>()); // presently hardcoding the index, inelegant but makes using FromJsonOverwrite very easy. TODO: better way?
