+++ Assets/Resources/Scripts/Map/Tiles/TileChars.cs
@@ -57,7 +57,7 @@ public class TileChars : MonoBehaviour
         {
             if(_boardStats == null)
             {
-                _boardStats = GetComponent<Tile>().board.GetComponent<BoardStats>();
+                _boardStats = GetComponent<TileLocation>().board.GetComponent<BoardStats>();
             }
             return _boardStats;
         }
