+++ Assets/Resources/Scripts/Tests/BoardStatsTestSuite.cs
@@ -23,10 +23,10 @@ public class BoardStatsTestSuite
 
         Vector2[] Edges = new Vector2[]
         {
-            Vector2.zero,
-            new Vector2(0, 5),
-            new Vector2(5, 5),
-            new Vector2(5, 0)
+            new Vector2(-.5f, -.5f),
+            new Vector2(-.5f, 4.5f),
+            new Vector2(4.5f, 4.5f),
+            new Vector2(4.5f, -.5f)
         };
 
         Assert.AreEqual(Edges[0], TestBoardStats.BoardEdges.BottomLeft);
